Extract account merge transaction
Amp-Thread-ID: https://ampcode.com/threads/T-01a05ac9-216a-7455-9050-5de42ba4d2a8
